Move S2ST submodules to separate files
