+++ fairseq/models/speech_to_speech/__init__.py
@@ -3,6 +3,5 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .modules import *  # noqa
 from .s2s_conformer import *  # noqa
 from .s2s_transformer import *  # noqa
